util folder modularized
